Dynamic interface, score, and next pieces are all now complete.
I had to make changes to the way Ty wrote the next piece code because it
no longer worked with the new interface. The graphics context for the
upcoming pieces is not the same panel as where the Tetris game pieces
are being drawn. So I had to come up with a slightly different way to do
it. The logic was similar, but a lot of things had to be added to
support it.